requirements.txt
# Intelligent System Resource Optimizer & Hardware Scheduler
# Core Dependencies

psutil>=5.9.0
GPUtil>=1.4.0
pystray>=0.19.0
Pillow>=10.0.0
rich>=13.0.0
PyYAML>=6.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
